Rename to `segment-rails` to match rails conventions
