+++ Gemfile
@@ -1,4 +1,4 @@
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in segment_rails.gemspec
+# Specify your gem's dependencies in segment-rails.gemspec
 gemspec
