+++ Rakefile
@@ -1,1 +1,8 @@
 require "bundler/gem_tasks"
+
+require 'rake/testtask'
+
+
+Rake::TestTask.new do |t|
+  t.pattern = "test/**/test_*.rb"
+end
