+++ lib/segment-rails.rb
