+++ lib/segment-rails/version.rb
