+++ test/all.rb
@@ -1,0 +1,1 @@
+require_relative 'test_segment_rails'
