+++ test/minitest_helper.rb
@@ -1,4 +1,4 @@
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
-require 'segment_rails'
+require 'segment-rails'
 
 require 'minitest/autorun'
